연구실 운영 기능 보강·인건비 정산 정정·반응형 UI·주석 정리
구성원
- 시트 업로드 날짜(엑셀 직렬값) ISO 정규화, 목록 활성/비활성 필터·검색
- 비활성 구성원 조회는 교수·관리자만(서버 강제), 목록 컬럼 재배열

연구과제·실적
- 상태·기간 산정을 총 과제기간 → 해당 연도 기간 기준으로 통일(과제/예산/인건비/대시보드)
- 참여 구성원·담당자 후보를 기간 내 재직자로 제한
- 실적 다중 사사(1/n 지분 집계)·과제 성과 자동 카운트, 제목 클릭 상세 팝업(조회·수정·삭제)
- 실적 시트 사사 관대 처리(미등록·다중·미입력 허용)

인건비(funds)
- 확정 집행 재계산·멱등화로 중복·예산 누적 방지, 통합 확정 버튼
- 통합학생인건비 연도 누적(선택 연도까지), 편성 잔여 실시간·이중차감 제거

자원·대시보드
- 인프라 장비 종류별 색·클릭 팝업·랙 간격, 자산 목록 뷰포트 자동 페이지
- 대시보드 균등 과제 제외

플랫폼
- 반응형 오버플로우 전역 보정(표 카드 스크롤·모달 줄바꿈·그리드 min-width)
- projects list_projects에 admin 포함, 코드 주석·죽은 코드 정리

+++ backend/services/funds/app/models.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import uuid
 from datetime import date, datetime
 
-from sqlalchemy import JSON, Date, DateTime, Integer, String, Text, func
+from sqlalchemy import JSON, Date, DateTime, Float, Integer, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from labmate_common.db import Base
@@ -61,7 +61,7 @@ class Participation(OrgScoped, SoftDelete, Base):
     id: Mapped[str] = mapped_column(String(32), primary_key=True, default=_uuid)
     uid: Mapped[str] = mapped_column(String(32), index=True)
     project_id: Mapped[str] = mapped_column(String(32))
-    rate_pct: Mapped[int] = mapped_column(Integer, default=0)
+    rate_pct: Mapped[float] = mapped_column(Float, default=0)   # 참여율(%) — 소수 허용
     month: Mapped[str] = mapped_column(String(7), index=True)   # YYYY-MM
